Update README typos and syntax
@avi here is an update for the promises and async/await readme. Everything looks great. I added a note about explicitly returning the `resolve()` function (i.e. `return resolve()` ) in a promise, in the cases where you might not want code to run inside of the new Promise function after resolve is invoked. I think this is a great example, but is this the very first time async is taught? If so maybe a more simplified example of thinking about promises. I always use the example of:

I'm starting to cook dinner that will take a few hours, but I am missing a couple of ingredients. So I ask my roommate to go to the store and pick me up some rice and butter. My roommate says sure and heads out ("The roommate is making a promise of returning from the store"). Though I am waiting for my rice and butter there are still things I can be prepping and cooking for dinner that don't rely upon the rice and butter so I continue making dinner. Once my roommate arrives with the rice and butter (moving into the .then function callback) I then take the rice and butter and start making the risotto and complete my dinner.  

Might be something you can use somewhere in the lectures or a README. Students tend to grasp onto that concept for promises. If an additional lessons you go onto explain Promise rejections with the `.catch` function. I extend the story about how there was trouble for my roommate to make it to the store. Like a road closure, a storm, an accident, or the store simply didn't have the ingredients. It seems rejections are not part of this lab, and I think they are just additional fluff that would over bloat this specific README at the moment.